sub2/backend/requirements.txt
apiview==1.3.25
argon2-cffi==20.1.0
astropy==4.0.1.post1
attrs==20.1.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
branca==0.3.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorful==0.5.4
confuse==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.7
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-import-export==2.3.0
django-redis==4.12.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
flex==6.14.1
folium==0.10.1
graphviz==0.14.1
htmlmin==0.1.12
idna==2.8
ImageHash==4.1.0
importlib-metadata==1.7.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.14.0
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.1.0
llvmlite==0.34.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.1
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numba==0.51.1
numpy==1.17.3
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
pandas==0.25.3
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.1
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
pprintpp==0.4.0
prettyprinter==0.18.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pydot3==1.0.9
Pygments==2.6.1
pyparsing==2.4.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.13.0
sqlparse==0.3.0
strict-rfc3339==0.7
tablib==2.0.0
tangled-up-in-unicode==0.0.6
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
validate-email==1.3
visions==0.4.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
